Added Details plugin to embedded, decoupled details logic from maps
