+++ web/client/epics/__tests__/maps-test.js
@@ -8,21 +8,14 @@
 
 import expect from 'expect';
 
-import configureMockStore from 'redux-mock-store';
-import { createEpicMiddleware, combineEpics } from 'redux-observable';
 import { CALL_HISTORY_METHOD } from 'connected-react-router';
 
 import {
     MAPS_LIST_LOADING,
     MAPS_LIST_LOADED,
     MAPS_LIST_LOAD_ERROR,
-    CLOSE_DETAILS_PANEL,
-    closeDetailsPanel,
     loadMaps,
     MAPS_GET_MAP_RESOURCES_BY_CATEGORY,
-    openDetailsPanel,
-    UPDATE_DETAILS,
-    DETAILS_LOADED,
     getMapResourcesByCategory,
     MAP_DELETING,
     MAP_DELETED,
@@ -37,27 +30,20 @@ import {
     LOAD_CONTEXTS
 } from '../../actions/maps';
 
-import { mapInfoLoaded, MAP_SAVED, LOAD_MAP_INFO, MAP_CONFIG_LOADED } from '../../actions/config';
+import { MAP_SAVED, LOAD_MAP_INFO, MAP_CONFIG_LOADED } from '../../actions/config';
 import { SHOW_NOTIFICATION } from '../../actions/notifications';
 import { TOGGLE_CONTROL, SET_CONTROL_PROPERTY } from '../../actions/controls';
-import { CLOSE_FEATURE_GRID } from '../../actions/featuregrid';
 import { loginSuccess, logout } from '../../actions/security';
 
 import {
     loadMapsEpic,
     getMapsResourcesByCategoryEpic,
-    closeDetailsPanelEpic,
-    fetchDataForDetailsPanel,
     deleteMapAndAssociatedResourcesEpic,
     mapsSetupFilterOnLogin,
-    storeDetailsInfoEpic,
     mapSaveMapResourceEpic,
     reloadMapsEpic
 } from '../maps';
 
-const rootEpic = combineEpics(closeDetailsPanelEpic);
-const epicMiddleware = createEpicMiddleware(rootEpic);
-const mockStore = configureMockStore([epicMiddleware]);
 import { testEpic, addTimeoutEpic, TEST_TIMEOUT } from './epicTestUtils';
 import axios from '../../libs/ajax';
 import MockAdapter from 'axios-mock-adapter';
@@ -80,36 +66,14 @@ const locale = {
     }
 };
 const mapId = 1;
-const mapId2 = 2;
 const mapId8 = 8;
-const detailsText = "<p>details of this map</p>";
-const detailsUri = "data/2";
 let map1 = {
     id: mapId,
     name: "name"
 };
-let map2 = {
-    id: mapId2,
-    name: "name2"
-};
 let map8 = {
     id: mapId8,
     name: "name"
-};
-const mapsState = {
-    maps: {
-        results: [map1]
-    },
-    mapInitialConfig: {
-        mapId
-    },
-    map: {
-        present: {
-            info: {
-                details: encodeURIComponent(detailsUri)
-            }
-        }
-    }
 };
 
 // category is required to support details map
@@ -157,116 +121,17 @@ const testMap2 = {
     ]
 };
 
-const mapAttributesEmptyDetails = {
-    "AttributeList": {
-        "Attribute": [
-            {
-                "name": "details",
-                "type": "STRING",
-                "value": EMPTY_RESOURCE_VALUE
-            }
-        ]
-    }
-};
-const mapAttributesWithoutDetails = {
-    "AttributeList": {
-        "Attribute": []
-    }
-};
-
 describe('maps Epics', () => {
     const oldGetDefaults = ConfigUtils.getDefaults;
-    let store;
     beforeEach(() => {
-        store = mockStore();
         ConfigUtils.getDefaults = () => ({
             geoStoreUrl: baseUrl
         });
     });
 
     afterEach(() => {
-        epicMiddleware.replaceEpic(rootEpic);
         ConfigUtils.getDefaults = oldGetDefaults;
     });
-
-    it('test closeDetailsPanel', (done) => {
-
-        store.dispatch(closeDetailsPanel());
-
-        setTimeout( () => {
-            try {
-                const actions = store.getActions();
-                expect(actions.length).toBe(2);
-                expect(actions[0].type).toBe(CLOSE_DETAILS_PANEL);
-                expect(actions[1].type).toBe(TOGGLE_CONTROL);
-            } catch (e) {
-                done(e);
-            }
-            done();
-        }, 50);
-
-    });
-    it('test fetchDataForDetailsPanel', (done) => {
-        map1.details = encodeURIComponent(detailsUri);
-        testEpic(addTimeoutEpic(fetchDataForDetailsPanel), 3, openDetailsPanel(), actions => {
-            expect(actions.length).toBe(3);
-            actions.map((action) => {
-                switch (action.type) {
-                case TOGGLE_CONTROL:
-                    expect(action.control).toBe("details");
-                    expect(action.property).toBe("enabled");
-                    break;
-                case CLOSE_FEATURE_GRID:
-                    expect(action.type).toBe(CLOSE_FEATURE_GRID);
-                    break;
-                case UPDATE_DETAILS:
-                    expect(action.detailsText.indexOf(detailsText)).toNotBe(-1);
-                    break;
-                default:
-                    expect(true).toBe(false);
-                }
-            });
-            done();
-        }, mapsState);
-    });
-    it('test fetchDataForDetailsPanel with Error', (done) => {
-        testEpic(addTimeoutEpic(fetchDataForDetailsPanel), 2, openDetailsPanel(), actions => {
-            expect(actions.length).toBe(2);
-            actions.map((action) => {
-                switch (action.type) {
-                case TOGGLE_CONTROL:
-                    expect(action.control).toBe("details");
-                    expect(action.property).toBe("enabled");
-                    break;
-                case SHOW_NOTIFICATION:
-                    expect(action.message).toBe("maps.feedback.errorFetchingDetailsOfMap");
-                    break;
-                default:
-                    expect(true).toBe(false);
-                }
-            });
-            done();
-        }, {
-            locale: {
-                messages: {
-                    maps: {
-                        feedback: {
-                            errorFetchingDetailsOfMap: "maps.feedback.errorFetchingDetailsOfMap"
-                        }
-                    }
-                }
-            },
-            mapInitialConfig: {
-                mapId
-            },
-            map: {
-                present: {
-                    info: {}
-                }
-            }
-        });
-    });
-
     it('test deleteMapAndAssociatedResourcesEpic, with map, details, thumbnail errors', (done) => {
         map1.thumbnail = "wronguri/data/5/";
         map1.details = "wronguri/data/6/";
@@ -417,48 +282,6 @@ describe('maps Epics', () => {
                 details: "wrong/uri/4"
             }
         });
-    });
-    it('test storeDetailsInfoEpic', (done) => {
-        testEpic(addTimeoutEpic(storeDetailsInfoEpic), 1, mapInfoLoaded(map2, mapId2), actions => {
-            expect(actions.length).toBe(1);
-            actions.map((action) => {
-                switch (action.type) {
-                case DETAILS_LOADED:
-                    expect(action.mapId).toBe(mapId2);
-                    expect(action.detailsUri).toBe("rest%2Fgeostore%2Fdata%2F3983%2Fraw%3Fdecode%3Ddatauri");
-                    break;
-                default:
-                    expect(true).toBe(false);
-                }
-            });
-            done();
-        }, {mapInitialConfig: {
-            "mapId": mapId2
-        }});
-    });
-    it('test storeDetailsInfoEpic when api returns NODATA value', (done) => {
-        const mock = new MockAdapter(axios);
-        mock.onGet().reply(200, mapAttributesEmptyDetails);
-        testEpic(addTimeoutEpic(storeDetailsInfoEpic), 1, mapInfoLoaded(map2, mapId2), actions => {
-            expect(actions.length).toBe(1);
-            actions.map((action) => expect(action.type).toBe(TEST_TIMEOUT));
-            mock.restore();
-            done();
-        }, {mapInitialConfig: {
-            "mapId": mapId2
-        }});
-    });
-    it('test storeDetailsInfoEpic when api doesnt return details', (done) => {
-        const mock = new MockAdapter(axios);
-        mock.onGet().reply(200, mapAttributesWithoutDetails);
-        testEpic(addTimeoutEpic(storeDetailsInfoEpic), 1, mapInfoLoaded(map2, mapId2), actions => {
-            expect(actions.length).toBe(1);
-            actions.map((action) => expect(action.type).toBe(TEST_TIMEOUT));
-            mock.restore();
-            done();
-        }, {mapInitialConfig: {
-            "mapId": mapId2
-        }});
     });
     it('it test loadMapsEpic when search text is a special character', (done) => {
         const searchText = 'tes/t\?:;@=&';
