+++ web/client/selectors/details.js
@@ -1,0 +1,9 @@
+/*
+* Copyright 2020, GeoSolutions Sas.
+* All rights reserved.
+*
+* This source code is licensed under the BSD-style license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+export const detailsTextSelector = state => state?.details?.detailsText;
